requirements.txt
gradio
langchain
langchain-community

pyttsx3

python-dotenv
requests
pypdf
langchain-groq

huggingface_hub
chromadb
langchain-chroma
sentence-transformers
mediapipe
opencv-python
langchain-huggingface
